==> Chapter_5/1. Comp w other myeloid data.Rmd <==
---
title: "Comparisons w other myeloid data"
---

#load sce & directory
```{r}
sce <- readRDS("/Users/reganhamel/Documents/PhD/PhD Dissertation/Chapter 5/scRNAseq/sce.RDS")

#the changes below were made to the BC sce and then saved into a new Ch5 version
#sce$cell_type[sce$cell_type == "Monoctye"] <- 'Monocyte'
#sce$cell_type <- factor(sce$cell_type, levels = c("Microglia", "Macrophage", "Monocyte", "Dendritic_Cells", "Neutrophil", "CAMs", "TCell"))
#sce$stage <- factor(sce$stage, levels = c("Ctrl", 1, 2, 3, 10, 21))
#colnames(sce) <- make.unique(colnames(sce), sep = "X")

path <- "/Users/reganhamel/Documents/PhD/PhD Dissertation/Chapter 5/scRNAseq/1. Comparing w lit/"
```

# Fig 2a, MG downregulate canonical marker genes. Seurat dot plot
```{r}
library(Seurat)
library(ggplot2)


# extract the MG and convert the sce to a Seurat object
seur <- as.Seurat(sce[,sce$cell_type == "Microglia"], data="logcounts", assay="logcounts")

MG_genes <- c("Olfml3","Sparc", "Trem2", "C1qa", "C1qb",  "Ctss","Csf1r", "Hexb","Gpr34", "P2ry12","Plxdc2", "Sall1", "Siglech","Cx3cr1","Tmem119","Serpine2")

DotPlot(seur,
        features=MG_genes,
        group.by = "stage",
        cols = "RdBu",
        scale =F,
        assay="logcounts") + theme(axis.text.x = element_text(angle = 45, hjust = 1)) 

ggsave(paste0(path,"/Fig 2a - MG markers-legend.png"), width = 8, height = 20)
```



## Heatmap
```{r}
sce.hm <- sce
# add an X to the end of the second instance of each duplicated barcode to give it a unique name (the heatmap code requires unique barcodes)
names <- make.unique(colnames(sce.hm), sep="X")

# rename the barcodes in the sce object of duplicated cell names
colnames(sce.hm) <- names
```


Reorder and label the cell types
```{r}
#Set the gene lists of interest
MG <- c("P2ry12","Plxdc2", "Siglech", "Sparc","Serpine2")
names(MG) <- "MG"

MG_MCd <- c("Tmem119", "Gpr34", "Olfml3","Fcrls")
names(MG_MCd) <- "MG_MCd"

Myel <- c("Aif1", "Hexb", "C1qa", "Trem2")
names(Myel) <- "Myel"

YFP <- c("S100a11", "Ms4a7", "Chil3", "Mgst1", "Ccl7", "Arg1")
names(YFP) <- "Infiltrating"


#order the factors
sce.hm$cells_cond <- paste0(sce$cell_type, "-", sce$stage)

sce.hm$cells_cond <- droplevels(factor(sce.hm$cells_cond))

sce.hm$cells_cond <- factor(sce.hm$cells_cond, levels = c("Microglia-Ctrl", "Microglia-1", "Microglia-2","Microglia-3", "Microglia-10","Microglia-21", "Macrophage-Ctrl", "Macrophage-1", "Macrophage-2",  "Macrophage-3", "Macrophage-10", "Macrophage-21","Monocyte-Ctrl", "Monocyte-1", "Monocyte-2", "Monocyte-3", "Monocyte-10", "Monocyte-21","Dendritic_Cells-Ctrl", "Dendritic_Cells-1", "Dendritic_Cells-2", "Dendritic_Cells-3", "Dendritic_Cells-10", "Dendritic_Cells-21", "Neutrophil-1", "Neutrophil-2", "Neutrophil-3",  "Neutrophil-10", "CAMs-Ctrl", "CAMs-1", "CAMs-3", "CAMs-2", "CAMs-10", "CAMs-21",  "TCell-Ctrl", "TCell-1", "TCell-2", "TCell-3", "TCell-10", "TCell-21"))
```

Generate and save the heatmap
```{r}
list <- c(MG, MG_MCd, Myel, YFP)

plotHeatmap(sce.hm, features=list, exprs_values="logcounts", cellheight=5, cellwidth=0.01, cluster_cols=F, cluster_rows=F, columns=order(sce.hm$cells_cond), colour_columns_by=c("cell_type", "stage"), file=paste0(path, "HM MG.png"), fontsize=5, show_rownames=T, center=F, symmetric=F)
```


#bar chart to compare smFISH
Generate the area plot (using sce_comp)
```{r}
gene <- "Ms4a7"
fl <- "Infiltrating"
#sce_comp <- sce[,sce$cell_type == "Microglia" | sce$cell_type == 'CAMs']#  & sce$cell_type != "CAMs" ]
sce_comp <- sce[,sce$fl == fl]
sce_comp$gene <- counts(sce_comp)[gene,] > 0
temp <- c("#D53E4F", "#FF8D59","#E6F598","#1A9850","#3288BD", "#542788")

```

```{r}
library(dplyr)
#count the number of genes of a certain UMI per cluster
# set the x and y values for the plot
y <- sce_comp$gene
sce_comp$stage <- factor(sce_comp$stage, levels = c("Ctrl","0", "1", "2", "3", "10", "21"))
sce_comp$stage[sce_comp$stage == "Ctrl"] <- 0

x <- (factor(sce_comp$stage))

# create the df containing the x and y values
df <- as.data.frame(table(y, x))
colnames(df) <- c("y", "x", "Freq")

# create data
cluster <- as.numeric(levels(df$x)[df$x]) # x Axis
value <- as.numeric(df$Freq)             # y Axis
group <- df$y       # group, one shape per group
data <- data.frame(cluster, value, group)


# plot
# Compute percentages with dplyr
data <- data  %>%
  group_by(cluster, group) %>%
  summarise(n = sum(value)) %>%
  mutate(percentage = n / sum(n))

d <- data[data$group !=F,]

pdf(paste0(path, gene,"_", fl, ".pdf"), height = 5, width = 3)
barplot(height = (d$percentage)*100, space=0.1, names.arg = levels(droplevels(factor(sce_comp$stage))), ylab = "Fraction of Cells (%)", ylim = c(0,100), col=temp[-1], beside=T) 
dev.off()
```


#Heatmap of HEG
#calculate the most highly expressed genes
```{r}
library(scater)
path <- "/Users/reganhamel/Documents/PhD/PhD Dissertation/Chapter 5/scRNAseq/1. Comparing w lit/Apoe/"
sce.hm <- sce
# add an X to the end of the second instance of each duplicated barcode to give it a unique name (the heatmap code requires unique barcodes)
names <- make.unique(colnames(sce.hm), sep="X")
colnames(sce.hm) <- names # rename the barcodes in the sce object of duplicated cell names

genesCtrl <- sort(calculateAverage(sce[,sce$stage == "Ctrl"]), decreasing = T)

genesSCI <- sort(calculateAverage(sce[,sce$condition == "SCI"]), decreasing = T)
```

```{r}
list <- unique(names(c(genesCtrl[1:15], genesSCI[1:15])))

#Generate and save the heatmap
plotHeatmap(sce.hm, features=list, exprs_values="logcounts", cellheight=5, cellwidth=0.01, cluster_cols=F, cluster_rows=F, columns=order(sce.hm$cells_cond), colour_columns_by=c("cell_type", "stage"), file=paste0(path, "Top expressed.png"), fontsize=5, show_rownames=T, center=F, symmetric=F)
```


#Heatmap of DAMS
```{r}
#load the sce
path <- "/Users/reganhamel/Documents/PhD/PhD Dissertation/Chapter 5/scRNAseq/1. Comparing w lit/Apoe/"
sce.hm <- sce

source("/Users/reganhamel/Documents/PhD/scRNAseq Workflow/PhD_Dissertation/Chapter_5/Sources/Fxn - Prep Genes.R")

#Establish the DAM gene lists (from Keren-Shaul et al., 2017; Figure 6)

#homeostatic genes
hG <- c("Cst3, Ctsd, Csf1r, Ctss, Sparc, C1qa, C1qb, Hexb, Cx3cr1, P2ry12, Cd33, Tmem119")
# DAM phase 1 (Trem independent)
DAM1 <- c("Apoe, Tyrobp, Ctsb, Ctsd, B2m, Fth1, Lyz2") 

# DAM phase 2 (Trem dependent)
DAM2 <- c("Trem2, Axl, Cst7, Ctsl, Lpl, Cd9, Csf1, Ccl6, Itgax, Clec7a, Lilrb4a, Timp2")  

# separate the string
hG <- prepG(hG, sep = ", ")
DAM1 <- prepG(DAM1, sep = ", ")
DAM2 <- prepG(DAM2, sep = ", ")

list <- c(hG, DAM1, DAM2)

#keep only genes in HVGs
list[!(list %in% rownames(sce))]
list <- list[(list %in% rownames(sce))] #Lilrb4a" removed

#Generate and save the heatmap
plotHeatmap(sce.hm, features=list, exprs_values="logcounts", cellheight=5, cellwidth=0.01, cluster_cols=F, cluster_rows=F, columns=order(sce.hm$cells_cond), colour_columns_by=c("cell_type", "stage"), file=paste0(path, "DAMs.png"), fontsize=5, show_rownames=T, center=F, symmetric=F)
```

#M1/M2 heatmap
```{r}
path <- "/Users/reganhamel/Documents/PhD/PhD Dissertation/Chapter 5/scRNAseq/1. Comparing w lit/"
sce.hm <- sce

#macrophage identities from the Zhang 2015 review & Milich 2021
M1 <- c("Il1a, Il1b, Il6, Il12a, Il12b, Il27, Il23a, Fcgr3, Fcgr2b, Ccl2, Ccl1, Il12a, Cxcl2, Cxcl13, Csf2, Csf3, Cd86, Marco, Nos2, Stat1, Irf5, Jun, Peli1, Tnf, Tnfaip3, Nfkbiz, Socs3") #not in HVG:

M2 <- c("Arg1, Chil3, Mrc1, Retnla, Igf1, Il1rn, Stat6, Klf2, Irf4, Pparg, Ppara, Ppard, Clec10a, Rnase2a, Ccl8, Ccl17, Ccl22, Ccl24, Cdh1, Ear2, Pdcd1lg2, Socs2")

#M2b <- c("Il10, Il6, Vegfa, Igf1, Cd86, Tnf, Fcgr1")

#M2c <- c("Tgfb1, Mrc1, Cd163, Slamf1, Sphk1, Thbs1, Hmox1, Stat3")

#keep only genes in HVGs
list <- prepG(c(M2), sep = ", ")
list[!(list %in% rownames(sce))]
list <- list[(list %in% rownames(sce))] #Lilrb4a" removed

#Generate and save the heatmap
plotHeatmap(sce.hm, features=list, exprs_values="logcounts", cellheight=5, cellwidth=0.01, cluster_rows=T, cluster_cols=F, columns=order(sce.hm$cells_cond), colour_columns_by=c("cell_type", "stage"), file=paste0(path, "M1.png"), fontsize=5, show_rownames=T, center=F, symmetric=F)
```


#cell cycling
# Extended Data Fig 5f Heatmap of cell cycling genes

Set the genes for the heatmap
```{r}
path <- "/Users/reganhamel/Documents/PhD/PhD Dissertation/Chapter 5/scRNAseq/1. Comparing w lit/Cycling/"
colnames(sce) <- make.unique(colnames(sce), sep="X")

#  read in a list of cell cycle markers, from Tirosh et al., 2015
cc.genes <- readLines(con="/Users/reganhamel/Documents/PhD/scRNAseq Workflow/PhD_Dissertation/Chapter_5/Sources/regev_lab_cell_cycle_genes.txt")
cc.genes <- stringr::str_to_title(cc.genes)

# s phase vs g2-m phase genes for label reference
s.genes <- cc.genes[1:43]
g2m.genes <- cc.genes[44:97]

# load the function for preparing gene lists from source file
source("/Users/reganhamel/Documents/PhD/scRNAseq Workflow/PhD_Dissertation/Chapter_5/Sources/Fxn - Prep Genes.R")

# format the list of genes gfor the plotHeatmap function
input <- prepG(cc.genes, sep=",")

# remove any genes not in the sce
input <- match(input, rownames(sce))
input <- input[!is.na(input)]

```

Generate the heatmaps for each cell type of interest 
```{r}
library(scater)

# set the cell type
CT <- "Microglia" #"MG"
sce_hm <- sce[,sce$cell_type == CT]
ave.counts <- calculateAverage(sce_hm[input,], exprs_values="logcounts")
to.keep <- ave.counts >= 0.1
input2 <- input[to.keep]

# plot and save the heatmap
plotHeatmap(sce_hm, features=input2, exprs_values="logcounts", cluster_cols=F, cluster_rows=T, columns=order(sce_hm$cycl), 
            zlim=c(0,3), colour_columns_by=c("cycl"), file=paste0(path, "S-G2-M Phase Genes stage", CT, ".png"), fontsize=5, show_rownames=T, center=F, symmetric=F, cellheight=5, cellwidth=0.01)

#plotHeatmap(sce_hm, features=input, exprs_values="logcounts", cellheight=4, cellwidth=0.03,
# zlim=NULL, center=FALSE, symmetric=FALSE, cluster_cols=FALSE,cluster_rows=F,
# colour_columns_by=c("stage","Cluster"), columns=order(sce_hm$Cluster), main= paste("S-G2-M Phase Genes"), fontsize_row=5, fontsize_col=0.2, treeheight_col=40, file=paste0(path, "S-G2-M Phase Genes stage", CT, ".png"))  
```

#mki67 boxplot
Generate the boxplots for the genes of interest 
```{r}
library(ggplot2)
library(RColorBrewer)
library(ggpubr)
# set the genes of interest
i <- "Mki67"

x <- sce$stage
col <- droplevels(sce$cell_type)

logs <- as.numeric(logcounts(sce[i]))
df <- data.frame(Logs=logs, x=x, colour=col)

#add empty slots for Ctrl to keep the spacing
df.a <- data.frame(Logs=c(0,0,0, 0, 0), x=c("Ctrl", "Ctrl", "Ctrl", "10", "21"), colour=c("MCd", "DC", "NP", "NP", "NP"))
df <- rbind(df, df.a)
comparisons = list(c("Ctrl", "1"), c("Ctrl", "2"), c("Ctrl", "3"), c("Ctrl", "10"), c("Ctrl", "21"))

ggplot(data=df, aes(y=Logs, x=x, fill=colour)) + geom_boxplot() + ggtitle(i) + theme(plot.title=element_text(hjust=0.5)) + ylab("logounts") + xlab("dpi") + scale_fill_brewer(palette="Accent", direction=1) 

ggsave(paste0(path,"boxplot_", i, ".png"), scale=.75, height=5, width =5)
```

#looking at cyclins
D peaks at G1
E peaks at G1/S
A is expressed in S/G2
B peaks at G2/mitosis
```{r}
library(scater)
cyclin.genes <- grep("^Ccn[abde][0-9]$", rowData(sce)$SYMBOL)
cyclin.genes <- rownames(sce)[cyclin.genes]
cyclin.genes

plotHeatmap(sce, order_columns_by="Cluster", 
            cluster_rows=FALSE, features=sort(cyclin.genes))
```

```{r}
#use sce with all genes (not just HVGs)
cyc_sce <- readRDS("/Users/reganhamel/Documents/PhD/PhD Dissertation/Chapter 4/scRNAseq/4. BC/BC by strain plus Jan/sce_UMAP_noBC_H5_C10_filt.RDS")

set.seed(100)
library(scran)
mm.pairs <- readRDS(system.file("exdata", "mouse_cycle_markers.rds", 
                                package="scran"))

# Using Ensembl IDs to match up with the annotation in 'mm.pairs'.
assignments <- cyclone(cyc_sce, mm.pairs, gene.names=rowData(cyc_sce)$ENSEMBL)
saveRDS(assignments, paste0(path, "assignments.RDS"))
```

#plot assignment scores & label sce
```{r}
plot(assignments$score$G1, assignments$score$G2M,
     xlab="G1 score", ylab="G2/M score", pch=16)

cyc_sce$cycl <- assignments$phases

#add labels to sce and save
sce$cycl <- cyc_sce$cycl[match(colnames(sce), colnames(cyc_sce))]
sce$cycl[is.na(sce$cycl)] <- "G0"
saveRDS(sce, "/Users/reganhamel/Documents/PhD/PhD Dissertation/Chapter 5/scRNAseq/sce.RDS")
```
cplot
```{r}
table(sce$cycl, sce$Cluster)

plotUMAP(sce, colour_by="cycl")
```

```{r}

sceH <- readRDS("/Users/reganhamel/Documents/PhD/PhD Dissertation/Chapter 3/scRNAseq/2. BC/No Ctrl filtering/sce_UMAP_noBC.RDS")
sceH <- sceH[,sceH$stage == "HC"  | sceH$stage == "1" | sceH$stage == "2"]
```

# hammond and milich cycling 
```{r}
library(scater)

colnames(sce) <- make.unique(colnames(sce), sep = "X")

list <- c("Hist1h2bc", "Birc5", "H2afx", "Ube2c", "Ccnb2", "Cenpa", "Stmn1", "Tuba1b", "Pclaf", "Hmgb2", 'Tubb5',"Ran", "Cks1b") #Hammond et al and Milich et al

list[!(list %in% rownames(sce))]

sceH <- sce[, sce$cell_type == "Microglia"]
#Generate and save the heatmap
name <- "CtlvAcute short_w_Milich MG2"
plotHeatmap(sceH, features=list, exprs_values="logcounts", cluster_cols=F, cluster_rows=T, columns=order(sceH$cycl), 
            zlim=c(0,3), colour_columns_by=c("cycl", "stage"), file=paste0(path, name,".png"), fontsize=5, show_rownames=T, center=F, symmetric=F, cellheight=5, cellwidth=0.01)
```
#heatmap of cyclin expression

#looking at cyclins
D peaks at G1
E peaks at G1/S
A is expressed in S/G2
B peaks at G2/mitosis
Mki67 G1, S, G2
```{r}
library(scater)
cyclin.genes <- grep("^Ccn[abde][0-9]$", rowData(sce)$SYMBOL)
cyclin.genes <- rownames(sce)[cyclin.genes]
cyclin.genes

list <- c("Ccnd1", "Ccnd2", "Ccnd3","Ccne1","Ccne2", "Ccna2",  "Ccnb1",  "Ccnb2","Mki67")
#sce$cycl <- factor(sce$cycl, levels = c("G0", "G1", "S", "G2M"))
```

```{r}
scehm <- sce[,sce$cell_type == "Microglia"]
scehm$cycl <- droplevels(factor(scehm$cycl))

plotHeatmap(scehm, features=list, exprs_values="logcounts", cellheight=8, cellwidth=0.02,
            zlim=c(0,2), center=FALSE, symmetric=FALSE, cluster_cols=F,cluster_rows=F,
            colour_columns_by=c("cycl"), columns=order(scehm$cycl), fontsize_row=5, fontsize_col=0.2, treeheight_col=40, file=paste0(path, "S-G2-M Phase Genes stage MG.png"))  

```

```{r}
library(scater)
cby <- "Mki67"
plotUMAP(sce, colour_by=cby)
ggsave(filename = paste0(cby, ".png"), path = path, scale = 1, width = 8, height = 8,  dpi = 300, limitsize = TRUE)

```

#pearson's residuals
- http://www.sthda.com/english/wiki/chi-square-test-of-independence-in-r (how to interpret)
```{r}
library(RColorBrewer)
library(corrplot)

All <- (table(sce$cycl, sce$Cluster)) #repeat for cycl

chisq <- chisq.test(All)

# convert the z-scores (standardized residuals) to chi square values
chisq.val <- chisq$stdres * chisq$stdres

# convert the chi square values to p values
chi.p <- pchisq(chisq.val, df=1, lower.tail=FALSE) 

# correct for multiple testing:
# compare the chi square values to the bonferonni adjusted p value
bf.p <- (0.05/(nrow(All)*ncol(All)))

chisq$stdres[1,1] <- -40
chisq$stdres[2,1] <- 45
 
pdf(paste0(path, "cluster stage lim corr.pdf"))
corrplot(chisq$stdres, p.mat=chi.p, is.cor = F, sig.level = bf.p, tl.col = "black", method="color", cl.length =10, insig = "blank")
dev.off()
```

# match stage quant between strains
```{r}
table(sce$strain, sce$stage)
SCE <- sce
Cre <- sce[,sce$strain == "Cre"]
Cx <- sce[,sce$strain == "Cx3"]

KEEP <- NULL
for (i in levels(sce$stage)) {
  CRE <- Cre[,Cre$stage == i]
  CX <- Cx[,Cx$stage == i]
  
  if (ncol(CRE) > ncol(CX)) {
    splitpro <- ncol(CX)/ncol(CRE)
    tr <- sample(ncol(CRE), round(ncol(CRE) * splitpro))
    keep <- colnames(CRE[,-tr ])
  } else {
    splitpro <- ncol(CRE)/ncol(CX)
    tr <- sample(ncol(CX), round(ncol(CX) * splitpro))
    keep <- colnames(CX[,-tr ])
  }
  
  KEEP <- c(KEEP, keep)
}

length(KEEP)

SCE <- SCE[,!(colnames(sce) %in% KEEP)] 

table(SCE$strain, SCE$stage)
table(SCE$strain, SCE$cell_type)

```


Generate the area plot (using sce_comp)
```{r}
library(tibble)
library(ggplot2)
library(plyr)
CT <- "Microglia"
sce_comp <- sce[,sce$cell_type == CT]
sce_comp$cycl <- factor(sce_comp$cycl, levels=c(rev(levels(sce_comp$cycl))))
# set HC to 0
sce_comp$stage <- factor(sce_comp$stage, levels = c("Ctrl","0", "1", "2", "3", "10", "21"))
sce_comp$stage[sce_comp$stage == "Ctrl"] <- 0

# set the x and y values for the plot
y <- sce_comp$cycl
x <- droplevels(factor(sce_comp$stage))

# create the df containing the x and y values
df <- as.data.frame(table(y, x))
colnames(df) <- c("y", "x", "Freq")

# create data
time <- as.numeric(levels(df$x)[df$x]) # x Axis
value <- as.numeric(df$Freq)             # y Axis
group <- df$y       # group, one shape per group
data <- data.frame(time, value, group)


# plot
# Compute percentages with dplyr
library(dplyr)
data <- data  %>%
  group_by(time, group) %>%
  summarise(n = sum(value)) %>%
  mutate(percentage = n / sum(n))

# Plot
# Plot
ggplot(data, aes(x=time, y=percentage, fill=group)) + 
  geom_area(alpha=0.7 , size=0.5, colour="black") + ylab("Fraction of Cells") + scale_x_continuous(name ="dpi", breaks=c(0,1,2, 3, 10, 21), labels=c("0"="Ctrl","1"="1","2"="2", "3"="3", "10"="10", "21"="21"))+ scale_fill_brewer(palette="Accent", direction=-1)


ggsave(paste0(path, "/Area Plot",CT, ".png"), height=5, width=6)
```

#male vs female
```{r}
sceA <- readRDS("/Users/reganhamel/Documents/PhD/PhD Dissertation/Chapter 5/scRNAseq/4. MG4 removed/sce.RDS")

dirA <- "/Users/reganhamel/Documents/PhD/PhD Dissertation/Chapter 5/scRNAseq/sexes/"
```

Plot the UMAP using the mouse strain/fluorophore of interest and grey out the cells not of interest
```{r}
library(scater)
library(RColorBrewer)
library(Seurat)
sce <- sceA

# name for labelling the UMAP
name <- "grey by sex"

# choose how to colour and shape the UMAP
#colours <- c(brewer.pal(nlevels(sceA$cell_type), "Accent"),"grey95")
colours <- c("#D53E4F", "#FF8D59","#E6F598","#1A9850","#3288BD", "#542788", "grey50")

# choose the level to determine the grey
subset <- "sex"

# choose the level to colour the non-grey cells
col.name <- "stage"

# automatically set parameters
level <- colData(sce)[subset][,1]

# load the sce and generate the UMAP for each timepoint
for (j in levels(factor(level))) {
  sce <- sceA
  
  #set the colour category for the grey
  grey <- level != j
  
  # rename levels to include "grey" for those not of interest
  col <- colData(sce)[col.name][,1] # set the level to colour the non-grey cells
  sce$grey <- factor(col, levels=c(levels(col), "grey"))
  sce$grey[grey] <- "grey"
  
  #sort the cells in order to have the grey cells last (and therefore underneath)
  sce <- sce[,rev(order(colnames(sce)[sce$grey]))]
  
  # generate the UMAP for the timepoint
  plotUMAP(sce, colour_by="grey",by_exprs_values="logcounts", point_size = 0) +  ggtitle(paste0(j))  + theme(plot.title=element_text(size=rel(2)))   + geom_point(aes(fill=colData(sce)$grey), colour="grey50", pch=21, stroke = .1, size =2, alpha = 0.6) + scale_fill_manual(values=colours, drop=F)#+ theme(legend.position="none")
  
  # name and save the plot to combine with the other timepoints  
  ggsave(filename=paste0(name," ", j, "egend.png"), path=dirA, scale=1, width=8, height=8, dpi=300, limitsize=TRUE) 
}

```

Plot just a single samples
```{r}
library(scater)
library(RColorBrewer)
library(Seurat)
sce <- sceA

# name for labelling the UMAP
name <- "grey by sex"

# choose how to colour and shape the UMAP
#colours <- c(brewer.pal(nlevels(sceA$cell_type), "Accent"),"grey95")
colours <- c("#F1A340", "#542788", "grey90")

# choose the level to determine the grey
subset <- "sampleID"

 
# choose the level to colour the non-grey cells
col.name <- "sex"


# load the sce and generate the UMAP for each timepoint

  sce <- sceA
  
  #set the colour category for the grey
  grey <- sce$sampleID == "SIGAF11" #| sce$sampleID == "SIGAG5" # | sce$sampleID == "SIGAE11" | sce$sampleID == "SIGAF11"
  
  # rename levels to include "grey" for those not of interest
  col <- colData(sce)[col.name][,1] # set the level to colour the non-grey cells
  sce$grey <- factor(col, levels=c(levels(factor(col)), "grey"))
  sce$grey[!grey] <- "grey"
  
  #sort the cells in order to have the grey cells last (and therefore underneath)
  sce <- sce[,rev(order(colnames(sce)[sce$grey]))]
  
  # generate the UMAP for the timepoint
  plotUMAP(sce, colour_by="grey",by_exprs_values="logcounts", point_size = 0)  + theme(plot.title=element_text(size=rel(2)))   + geom_point(aes(fill=colData(sce)$grey), colour="grey70", pch=21, stroke = .1, size =1, alpha = 0.6) + scale_fill_manual(values=colours, drop=F)
  
  # name and save the plot to combine with the other timepoints  
 ggsave(filename=paste0("SIGAF11.png"), path=dirA, scale=1, width=8, height=8, dpi=300, limitsize=TRUE) 


```
